test: "사용하지 않는 테스트 삭제 및 kotlin test를 위한 config 설정"
