+++ src/test/kotlin/com/tobe/healthy/lessonHistory/application/LessonHistoryServiceTest.kt
@@ -1,70 +1,86 @@
 package com.tobe.healthy.lessonHistory.application
 
-import com.querydsl.jpa.impl.JPAQueryFactory
 import com.tobe.healthy.config.error.CustomException
+import com.tobe.healthy.config.error.ErrorCode.LESSON_HISTORY_COMMENT_NOT_FOUND
 import com.tobe.healthy.config.error.ErrorCode.LESSON_HISTORY_NOT_FOUND
-import com.tobe.healthy.file.repository.FileRepository
-import com.tobe.healthy.lessonHistory.domain.dto.LessonHistoryCommandResult
-import com.tobe.healthy.lessonHistory.domain.dto.LessonHistoryUpdateCommand
 import com.tobe.healthy.lessonHistory.domain.entity.LessonHistory
 import com.tobe.healthy.lessonHistory.domain.entity.LessonHistoryComment
-import com.tobe.healthy.lessonHistory.domain.entity.QLessonHistory
 import com.tobe.healthy.lessonHistory.repository.LessonHistoryCommentRepository
 import com.tobe.healthy.lessonHistory.repository.LessonHistoryRepository
-import com.tobe.healthy.log
 import com.tobe.healthy.member.domain.entity.Member
 import com.tobe.healthy.member.domain.entity.MemberType.STUDENT
 import com.tobe.healthy.member.domain.entity.MemberType.TRAINER
 import com.tobe.healthy.schedule.domain.entity.ReservationStatus.COMPLETED
 import com.tobe.healthy.schedule.domain.entity.Schedule
+import io.kotest.core.spec.style.BehaviorSpec
+import io.kotest.matchers.shouldBe
 import jakarta.persistence.EntityManager
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
 import java.time.LocalTime
-import java.util.stream.Collectors
 
-@SpringBootTest
 @Transactional
-class LessonHistoryServiceTest @Autowired constructor(
+@SpringBootTest
+class LessonHistoryServiceTest(
     private val em: EntityManager,
     private val passwordEncoder: PasswordEncoder,
-    private val queryFactory: JPAQueryFactory,
     private val lessonHistoryRepository: LessonHistoryRepository,
-    private val lessonHistoryCommandRepository: LessonHistoryCommentRepository,
-    private val fileRepository: FileRepository,
-) {
+    private val lessonHistoryCommentRepository: LessonHistoryCommentRepository,
+) : BehaviorSpec({
 
-    @Autowired
-    private lateinit var lessonHistoryCommentRepository: LessonHistoryCommentRepository
-    lateinit var student: Member
-    lateinit var trainer: Member
-    lateinit var schedule: Schedule
-    lateinit var lessonHistory: LessonHistory
+    Given("회원가입을 하고") {
+        val student = createStudent(passwordEncoder.encode("pass123"))
+        val trainer = createTrainer(passwordEncoder.encode("pass123"))
+        em.persist(student)
+        em.persist(trainer)
+        When("게시글과 댓글을 등록했을 때") {
+            val schedule = createSchedule(trainer, student)
+            val lessonHistory = createLessonHistory("테스트 게시글 제목", "테스트 게시글 내용", trainer, student, schedule)
+            em.persist(schedule)
+            em.persist(lessonHistory)
+            val comment = createLessonHistoryComment(lessonHistory)
+            em.persist(comment)
+            em.flush()
+            em.clear()
 
-    @BeforeEach
-    fun `초기 데이터를 설정한다`() {
-        student = Member.builder()
-            .userId("trainer")
-            .email("trainer@gmail.com")
-            .password(passwordEncoder.encode("zxcvbnm11"))
-            .name("trainer")
-            .memberType(TRAINER)
-            .build()
+            Then("게시글이 정상적으로 등록되었는지 검증한다") {
+                val result = lessonHistoryRepository.findByIdOrNull(lessonHistory.id) ?: throw CustomException(LESSON_HISTORY_NOT_FOUND)
+                result.content shouldBe lessonHistory.content
+            }
 
-        trainer = Member.builder()
-            .userId("student")
-            .email("student@gmail.com")
-            .password(passwordEncoder.encode("zxcvbnm11"))
-            .name("student")
-            .memberType(STUDENT)
-            .build()
+            Then("댓글이 정상적으로 등록되었는지 검증한다") {
+                val result = lessonHistoryCommentRepository.findByIdOrNull(comment.id) ?: throw CustomException(LESSON_HISTORY_COMMENT_NOT_FOUND)
+                result.content = comment.content
+            }
+        }
+    }
+})
+{ companion object {
+        private fun createLessonHistoryComment(lessonHistory: LessonHistory): LessonHistoryComment {
+        return LessonHistoryComment(
+            order = 1,
+            content = "테스트용 댓글 등록",
+            writer = lessonHistory.student,
+            lessonHistory = lessonHistory)
+        }
 
-        schedule = Schedule.builder()
+        private fun createLessonHistory(title: String, content: String, trainer: Member, student: Member, schedule: Schedule) : LessonHistory {
+            return LessonHistory(
+                title,
+                content,
+                mutableListOf(),
+                mutableListOf(),
+                trainer,
+                student,
+                schedule
+            )
+        }
+
+        private fun createSchedule(trainer: Member, student: Member): Schedule {
+        return Schedule.builder()
             .lessonDt(LocalDate.of(2024, 4, 1))
             .lessonStartTime(LocalTime.of(10, 0))
             .lessonEndTime(LocalTime.of(11, 0))
@@ -73,127 +89,26 @@ class LessonHistoryServiceTest @Autowired constructor(
             .trainer(trainer)
             .applicant(student)
             .build()
-
-        lessonHistory = LessonHistory(
-            "title",
-            "content",
-            mutableListOf(),
-            mutableListOf(),
-            trainer,
-            student,
-            schedule)
-
-        em.persist(student)
-        em.persist(trainer)
-        em.persist(schedule)
-        em.persist(lessonHistory)
-    }
-
-    @Test
-    fun `게시글에 댓글을 등록한다`() {
-        val order = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment = getLessonHistoryComment(order, "123213", student)
-        em.persist(lessonHistoryComment)
-    }
-
-    @Test
-    fun `게시글에 댓글을 두개 등록한다`() {
-        val order = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment = getLessonHistoryComment(order, "댓글 내용입니다!!", student)
-        em.persist(lessonHistoryComment)
-
-        val order2 = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment2 = getLessonHistoryComment(order2, "댓글 내용입니다!!(2)", trainer)
-        em.persist(lessonHistoryComment2)
-    }
-
-    @Test
-    fun `게시글에 대댓글을 등록한다`() {
-        val order = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment = getLessonHistoryComment(order, "댓글 내용입니다!!", student)
-        em.persist(lessonHistoryComment)
-
-        val order2 = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!, lessonHistoryComment.id!!) ?: 1
-        val lessonHistoryComment2 = getLessonHistoryComment(order2, "대댓글 작성!!", trainer, lessonHistoryComment)
-        em.persist(lessonHistoryComment2)
-
-        val order3 = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!, lessonHistoryComment.id!!) ?: 1
-        val lessonHistoryComment3 = getLessonHistoryComment(order3, "대댓글 작성 두번 째!!", student, lessonHistoryComment)
-        em.persist(lessonHistoryComment3)
-    }
-
-    @Test
-    fun `게시글을 조회한다`() {
-        val lessonHistory = queryFactory
-            .selectDistinct(QLessonHistory.lessonHistory)
-            .from(QLessonHistory.lessonHistory)
-            .leftJoin(QLessonHistory.lessonHistory.lessonHistoryComment).fetchJoin()
-            .fetch()
-
-        val results = lessonHistory.stream().map { entity -> LessonHistoryCommandResult.from(entity) }.collect(Collectors.toList())
-
-        for (result in results) {
-            log.info { "result => ${result}" }
         }
-    }
 
-    @Test
-    fun `게시글을 삭제한다`() {
-        lessonHistoryRepository.delete(lessonHistory)
-    }
-
-    @Test
-    fun `댓글을 삭제한다`() {
-        val order = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment = getLessonHistoryComment(order, "댓글 내용입니다!!", student)
-        em.persist(lessonHistoryComment)
-        lessonHistoryCommandRepository.delete(lessonHistoryComment)
-    }
-
-    @Test
-    fun `게시글을 수정한다`() {
-        val request = LessonHistoryUpdateCommand(
-            title = "이걸로 바꿔줘",
-            content = "내용은 이걸로 바꿔줘"
-        )
-        val entity = lessonHistoryRepository.findById(lessonHistory.id!!).orElseThrow()
-        entity.updateLessonHistory(request.title, request.content)
-    }
-
-    @Test
-    fun `댓글을 수정한다`() {
-        val order = lessonHistoryCommandRepository.findTopComment(lessonHistory.id!!) ?: 1
-        val lessonHistoryComment = getLessonHistoryComment(order, "댓글 내용입니다!!", student)
-        em.persist(lessonHistoryComment)
-        val entity = lessonHistoryCommentRepository.findById(lessonHistoryComment.id!!).orElseThrow()
-        entity.updateLessonHistoryComment("댓글 내용은 이걸로 변경")
-        em.flush()
-    }
-
-    @Test
-    fun `첨부 파일을 조회한다`() {
-        val entity = queryFactory
-            .selectDistinct(QLessonHistory.lessonHistory)
-            .from(QLessonHistory.lessonHistory)
-            .innerJoin(QLessonHistory.lessonHistory.trainer).fetchJoin()
-            .innerJoin(QLessonHistory.lessonHistory.student).fetchJoin()
-            .innerJoin(QLessonHistory.lessonHistory.schedule).fetchJoin()
-            .leftJoin(QLessonHistory.lessonHistory.lessonHistoryComment).fetchJoin()
-            .fetch() ?: throw CustomException(LESSON_HISTORY_NOT_FOUND)
-
-        val results = entity.stream().map { e -> LessonHistoryCommandResult.from(e) }
-        for (result in results) {
-            log.info { "result => $result" }
+        private fun createStudent(password: String): Member {
+            return Member.builder()
+                .userId("student")
+                .email("student@gmail.com")
+                .password(password)
+                .name("student")
+                .memberType(STUDENT)
+                .build()
         }
-    }
 
-    private fun getLessonHistoryComment(order: Int, content: String, writer: Member, parentId: LessonHistoryComment? = null): LessonHistoryComment {
-       return LessonHistoryComment(
-            order = order,
-            content = content,
-            writer = writer,
-            lessonHistory = lessonHistory,
-            parentId = parentId
-        )
+        private fun createTrainer(password: String): Member {
+            return Member.builder()
+                .userId("trainer")
+                .email("trainer@gmail.com")
+                .password(password)
+                .name("trainer")
+                .memberType(TRAINER)
+                .build()
+        }
     }
 }
